all the missing vector tests

+++ src/main/java/de/bixilon/kmath/vec/vec2/i/MVec2i.kt
@@ -75,7 +75,7 @@ value class MVec2i(
     inline operator fun inc() = this + 1
     inline operator fun dec() = this - 1
 
-    inline fun length() = sqrt(length2().toDouble()).toInt()
+    inline fun length() = sqrt(length2().toDouble())
     inline fun length2() = x * x + y * y
